Events -> Sessions
Also made it to save each POST request as a seperate document to avoid
race conditions in the database when multiple lambdas activate at once
and make multiple documents. The new plan is to use an aggregator on
data pull from the database to gather all the events from a session in a
sorted manner.